feat: fill in `Report` and `GraphicalReportHandler`
Renamed `Error` to `Report` = to conform with [`std::error::Report`](https://doc.rust-lang.org/std/error/struct.Report.html).

`Error` is a trait, a `Report` is an error reporter.

As for the implementation, miette uses a bunch of indirections with unsafe implementations to store the `Diagnostic` trait for the `Report` struct.
I don't understand what's doing so I went with the simplest solution -
copy `std::error::Report` and save the `Diagnostic` trait into the `Report` struct.

I have a feeling that miette and eyerish's implementation predates what
we can do with Rust as of today, because miette's implementation looks
like a vtable for dynamic dispatch, which is exactly what the current implementation is doing.

Let's see how this pans out.